long

+++ 00_ProjetProxyBanque/src/main/java/fr/adaming/dao/IConseillerDao.java
@@ -24,7 +24,7 @@ public interface IConseillerDao {
 	
 	public Conseiller getConseillerDao(Conseiller conseiller);
 	
-	public int verifConseillerDao(Conseiller conseiller);
+	public Long verifConseillerDao(Conseiller conseiller);
 	
 
 }
